@@ -14,45 +14,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.1",
-    "@angular/common": "~10.0.1",
-    "@angular/compiler": "~10.0.1",
-    "@angular/core": "~10.0.1",
-    "@angular/forms": "~10.0.1",
-    "@angular/platform-browser": "~10.0.1",
-    "@angular/platform-browser-dynamic": "~10.0.1",
-    "@angular/router": "~10.0.1",
-    "@ngrx/data": "^9.0.0",
-    "@ngrx/effects": "^9.0.0",
-    "@ngrx/entity": "^9.0.0",
-    "@ngrx/store": "^9.0.0",
-    "@ngrx/store-devtools": "^9.0.0",
+    "@angular/animations": "^16.0.3",
+    "@angular/common": "^16.0.3",
+    "@angular/compiler": "^16.0.3",
+    "@angular/core": "^16.0.3",
+    "@angular/forms": "^16.0.3",
+    "@angular/platform-browser": "^16.0.3",
+    "@angular/platform-browser-dynamic": "^16.0.3",
+    "@angular/router": "^16.0.3",
+    "@ngrx/data": "^16.0.0",
+    "@ngrx/effects": "^16.0.0",
+    "@ngrx/entity": "^16.0.0",
+    "@ngrx/store": "^16.0.0",
+    "@ngrx/store-devtools": "^16.0.0",
     "font-awesome": "^4.7.0",
     "rxjs": "~6.5.5",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.0",
-    "@angular/cli": "~10.0.0",
-    "@angular/compiler-cli": "~10.0.1",
-    "@angular/language-service": "~10.0.1",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "^16.0.3",
+    "@angular/cli": "^16.0.3",
+    "@angular/compiler-cli": "^16.0.3",
+    "@angular/language-service": "^16.0.3",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "bulma": "^0.8.1",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^6.0.0",
     "concurrently": "^5.1.0",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "json-server": "^0.16.1",
-    "karma": "~5.0.0",
+    "karma": "~6.4.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~3.3.0",
+    "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~4.9.5"
   }
 }
