@@ -5,43 +5,43 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "11.0.0",
-    "@angular/cdk": "11.0.0",
-    "@angular/common": "11.0.0",
-    "@angular/compiler": "11.0.0",
-    "@angular/core": "11.0.0",
-    "@angular/forms": "11.0.0",
-    "@angular/material": "11.0.0",
-    "@angular/platform-browser": "11.0.0",
-    "@angular/platform-browser-dynamic": "11.0.0",
-    "@angular/router": "11.0.0",
-    "angular-in-memory-web-api": "^0.11.0",
+    "@angular/animations": "16.0.3",
+    "@angular/cdk": "16.0.2",
+    "@angular/common": "16.0.3",
+    "@angular/compiler": "16.0.3",
+    "@angular/core": "16.0.3",
+    "@angular/forms": "16.0.3",
+    "@angular/material": "16.0.2",
+    "@angular/platform-browser": "16.0.3",
+    "@angular/platform-browser-dynamic": "16.0.3",
+    "@angular/router": "16.0.3",
+    "angular-in-memory-web-api": "^0.16.0",
     "rxjs": "^6.6.3",
     "subsink": "^0.1.0",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.1",
-    "@angular/cli": "^11.0.1",
-    "@angular/compiler-cli": "^11.0.0",
-    "@angular/language-service": "11.0.0",
+    "@angular-devkit/build-angular": "^16.0.3",
+    "@angular/cli": "^16.0.3",
+    "@angular/compiler-cli": "^16.0.3",
+    "@angular/language-service": "16.0.3",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
-    "karma": "~5.0.0",
+    "karma": "~6.4.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "4.0.5"
+    "typescript": "4.9.5"
   }
 }
