@@ -2,10 +2,10 @@
   "name": "shared-lib",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^11.0.0",
-    "@angular/core": "^11.0.0"
+    "@angular/common": "^16.0.3",
+    "@angular/core": "^16.0.3"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": "^2.5.2"
   }
 }
