@@ -6,16 +6,20 @@ import {
   OnChanges,
   Output,
   ViewChild,
-  SimpleChanges
-} from "@angular/core";
+  SimpleChanges,
+} from '@angular/core';
 
-import { Villain } from "../../core";
-import { Validators, FormBuilder, FormGroup } from "@angular/forms";
+import { Villain } from '../../core';
+import {
+  Validators,
+  UntypedFormBuilder,
+  UntypedFormGroup,
+} from '@angular/forms';
 
 @Component({
-  selector: "app-villain-detail",
-  templateUrl: "./villain-detail.component.html",
-  styleUrls: ["./villain-detail.component.scss"]
+  selector: 'app-villain-detail',
+  templateUrl: './villain-detail.component.html',
+  styleUrls: ['./villain-detail.component.scss'],
 })
 export class VillainDetailComponent implements OnChanges {
   @Input() villain: Villain;
@@ -23,17 +27,17 @@ export class VillainDetailComponent implements OnChanges {
   @Output() add = new EventEmitter<Villain>();
   @Output() update = new EventEmitter<Villain>();
 
-  @ViewChild("name", { static: true }) nameElement: ElementRef;
+  @ViewChild('name', { static: true }) nameElement: ElementRef;
 
   addMode = false;
 
   form = this.fb.group({
     id: [],
-    name: ["", Validators.required],
-    saying: [""]
+    name: ['', Validators.required],
+    saying: [''],
   });
 
-  constructor(private fb: FormBuilder) {}
+  constructor(private fb: UntypedFormBuilder) {}
 
   ngOnChanges(changes: SimpleChanges) {
     this.setFocus();
@@ -46,7 +50,7 @@ export class VillainDetailComponent implements OnChanges {
     }
   }
 
-  addVillain(form: FormGroup) {
+  addVillain(form: UntypedFormGroup) {
     const { value, valid, touched } = form;
     if (touched && valid) {
       this.add.emit({ ...this.villain, ...value });
@@ -58,7 +62,7 @@ export class VillainDetailComponent implements OnChanges {
     this.unselect.emit();
   }
 
-  saveVillain(form: FormGroup) {
+  saveVillain(form: UntypedFormGroup) {
     if (this.addMode) {
       this.addVillain(form);
     } else {
@@ -70,7 +74,7 @@ export class VillainDetailComponent implements OnChanges {
     this.nameElement.nativeElement.focus();
   }
 
-  updateVillain(form: FormGroup) {
+  updateVillain(form: UntypedFormGroup) {
     const { value, valid, touched } = form;
     if (touched && valid) {
       this.update.emit({ ...this.villain, ...value });
