@@ -1,16 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CommunicationRoutingModule } from './communication-routing.module';
+import { routes, components } from './routes';
 import { CustomersListComponent } from './customers-list/customers-list.component';
 import { CustomerDetailsComponent } from './customer-details/customer-details.component';
-
-
+import { RouterModule } from '@angular/router';
 
 @NgModule({
-  imports: [
-    CommonModule,
-    CommunicationRoutingModule  
-  ],
-  declarations: [ CommunicationRoutingModule.components, CustomersListComponent, CustomerDetailsComponent ]
+  imports: [CommonModule, RouterModule.forRoot(routes)],
+  declarations: [components, CustomersListComponent, CustomerDetailsComponent],
 })
-export class CommunicationModule { }
+export class CommunicationModule {}
