@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { CustomersRoutingModule } from './customers-routing.module';
+import { components, routes } from './routes';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
-    imports: [ CommonModule, CustomersRoutingModule ],
-    declarations: [ CustomersRoutingModule.components ]
+  imports: [CommonModule, RouterModule.forChild(routes)],
+  declarations: [components],
 })
-export class CustomersModule { }
+export class CustomersModule {}
