@@ -1,11 +1,14 @@
 import { Injectable } from '@angular/core';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatLegacySnackBar as MatSnackBar } from '@angular/material/legacy-snack-bar';
 
 @Injectable({ providedIn: 'root' })
 export class ToastService {
   constructor(public snackBar: MatSnackBar) {}
 
   openSnackBar(message: string, action: string) {
-    setTimeout(() => this.snackBar.open(message, action, { duration: 2000 }), 0);
+    setTimeout(
+      () => this.snackBar.open(message, action, { duration: 2000 }),
+      0
+    );
   }
 }
