@@ -3,10 +3,11 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 
-import { AppRoutingModule } from './app-routing.module';
+import { routes } from './routes';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core/core.module';
 import { AppStoreModule } from './store/app-store.module';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   imports: [
@@ -14,10 +15,10 @@ import { AppStoreModule } from './store/app-store.module';
     BrowserAnimationsModule,
     CoreModule,
     HttpClientModule,
-    AppRoutingModule,
-    AppStoreModule
+    RouterModule.forRoot(routes),
+    AppStoreModule,
   ],
   declarations: [AppComponent],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule {}
