@@ -5,10 +5,15 @@ import { RouterModule } from '@angular/router';
 
 import { SharedModule } from '../shared/shared.module';
 import { OrdersComponent } from './orders.component';
-import { OrdersRoutingModule } from './orders-routing.module';
+import { routes } from './routes';
 
 @NgModule({
-    imports: [ CommonModule, FormsModule, SharedModule, OrdersRoutingModule ],
-    declarations: [ OrdersComponent ]
+  imports: [
+    CommonModule,
+    FormsModule,
+    SharedModule,
+    RouterModule.forChild(routes),
+  ],
+  declarations: [OrdersComponent],
 })
-export class OrdersModule { }
+export class OrdersModule {}
