@@ -4,25 +4,36 @@ import { StoreModule } from '@ngrx/store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { environment } from '../../environments/environment';
 
-import { DefaultDataServiceConfig, EntityDataModule } from '@ngrx/data';
-
+import {
+  DefaultDataServiceConfig,
+  EntityDataModule,
+  EntityDataModuleConfig,
+} from '@ngrx/data';
 
 const apiRoot = environment.apiUrlBase + '/';
 const defaultDataServiceConfig: DefaultDataServiceConfig = {
   root: apiRoot,
   entityHttpResourceUrls: {
-    Customer: { entityResourceUrl: apiRoot + 'customers/', collectionResourceUrl: apiRoot + 'customers/' },
-    Order: { entityResourceUrl: apiRoot + 'orders/', collectionResourceUrl: apiRoot + 'orders/' },
-  }
+    Customer: {
+      entityResourceUrl: apiRoot + 'customers/',
+      collectionResourceUrl: apiRoot + 'customers/',
+    },
+    Order: {
+      entityResourceUrl: apiRoot + 'orders/',
+      collectionResourceUrl: apiRoot + 'orders/',
+    },
+  },
 };
 
 @NgModule({
   imports: [
     StoreModule.forRoot({}),
     EffectsModule.forRoot([]),
-    EntityDataModule.forRoot(),
-    environment.production ? [] : StoreDevtoolsModule.instrument()
+    EntityDataModule.forRoot([] as EntityDataModuleConfig),
+    environment.production ? [] : StoreDevtoolsModule.instrument(),
   ],
-  providers: [ { provide: DefaultDataServiceConfig, useValue: defaultDataServiceConfig } ]
+  providers: [
+    { provide: DefaultDataServiceConfig, useValue: defaultDataServiceConfig },
+  ],
 })
 export class AppStoreModule {}
