@@ -1,14 +1,20 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 
 import { SharedModule } from '../shared/shared.module';
 import { OrdersComponent } from './orders.component';
-import { OrdersRoutingModule } from './orders-routing.module';
+
+const routes: Routes = [{ path: '', component: OrdersComponent }];
 
 @NgModule({
-    imports: [ CommonModule, FormsModule, SharedModule, OrdersRoutingModule ],
-    declarations: [ OrdersComponent ]
+  imports: [
+    CommonModule,
+    FormsModule,
+    SharedModule,
+    RouterModule.forChild(routes),
+  ],
+  declarations: [OrdersComponent],
 })
-export class OrdersModule { }
+export class OrdersModule {}
